Refactor extract log screen to its own module
